+++ tests/conftest.py
@@ -20,3 +20,4 @@ def user(app):
         'password': '5354',
         'client': client}
     return data
+
