--- tests/test_contact.py
@@ -1,4 +1,0 @@
-def test_app(app):
-    client = app.test_client()
-    resp = client.get('/contact')
-    assert resp.status_code == 200
